requirements.txt
bcrypt==4.3.0
dnspython==2.7.0
dotenv==0.9.9
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pymongo==4.11.3
pytest==8.3.5
python-dotenv==1.0.1
